Move conflict resolution to the separate folder
