+++ src/ggrc-client/js/models/conflict-resolution/conflict-resolution.js
